LayoutTests/imported/w3c/web-platform-tests/fs/FileSystemSyncAccessHandle-getSize.https.tentative.worker-expected.txt


FAIL test SyncAccessHandle.getSize after SyncAccessHandle.write assert_equals: expected (number) 0 but got (object) object "[object Promise]"
